requirements.txt
openai
python-dotenv
python-chess